+++ app/(feature)/form/hooks/index.ts
@@ -1,2 +1,1 @@
 export { useLeavingModal } from './useLeavingModal';
-export { usePostHelp } from './usePostHelp';
